Source: https://github.com/msupply-foundation/open-msupply/commit/c3be94187478c1727633c4f436563eed38752f0d

+++ server/graphql/demographic/src/mutations/insert_projection.rs
@@ -49,11 +49,11 @@ pub fn insert_demographic_projection(
 pub struct InsertDemographicProjectionInput {
     pub id: String,
     pub base_year: i32,
-    pub year_1: Option<i32>,
-    pub year_2: Option<i32>,
-    pub year_3: Option<i32>,
-    pub year_4: Option<i32>,
-    pub year_5: Option<i32>,
+    pub year_1: Option<f64>,
+    pub year_2: Option<f64>,
+    pub year_3: Option<f64>,
+    pub year_4: Option<f64>,
+    pub year_5: Option<f64>,
 }
 
 impl From<InsertDemographicProjectionInput> for InsertDemographicProjection {
